rename user-lists dir to resource-lists
