+++ frontends/infinite-corridor/src/App.tsx
@@ -7,9 +7,9 @@ import DemoPage from "./pages/DemoPage"
 import FieldPage from "./pages/field-details/FieldPage"
 import FieldAdminApp from "./pages/field-details/FieldAdminApp"
 import LearningResourceDrawer from "./components/LearningResourceDrawer"
-import UserListsListingPage from "./pages/user-lists/UserListsListingPage"
-import UserListDetailsPage from "./pages/user-lists/UserListDetailsPage"
-import FavoritesPage from "./pages/user-lists/FavoritesPage"
+import UserListsListingPage from "./pages/resource-lists/UserListsListingPage"
+import UserListDetailsPage from "./pages/resource-lists/UserListDetailsPage"
+import FavoritesPage from "./pages/resource-lists/FavoritesPage"
 import * as urls from "./pages/urls"
 import { Route, Router, Switch } from "react-router"
 import { History } from "history"
