+++ frontends/infinite-corridor/src/pages/resource-lists/AddToListDialog.tsx
