+++ frontends/infinite-corridor/src/pages/resource-lists/FavoritesPage.test.tsx
