+++ frontends/infinite-corridor/src/pages/resource-lists/ItemsListing.tsx
