+++ frontends/infinite-corridor/src/pages/resource-lists/ManageListDialogs.tsx
