+++ frontends/infinite-corridor/src/pages/resource-lists/UpsertListDialog.tsx
