+++ frontends/infinite-corridor/src/pages/resource-lists/UserListDetailsPage.tsx
