+++ frontends/infinite-corridor/src/pages/resource-lists/UserListsListingPage.tsx
